Switch riff BuildTemplate to be cluster scoped
Knative supports both BuildTemplate and ClusterBuildTemplate resources.
The advantage of the ClusterBuildTemplate is that it can be maintained
and upgraded as part of riff.

Currently, the template is deployed in each namespace and there is no
mechanism to maintain the template.

This PR also splits out the build cache from the build template. The
cache and service account will still need to be created in each
namespace where builds will run.

There will be a coresponding PR in riff to pull in these changes.

Refs projectriff/riff#968