Parameterize builder and util images in build template
The builder template is used twice within the build template, as is the
util image. We can centralize the definition of the builder image by
using a parameter. In general, builds should not override the run or
builder image params. It is still useful to parameterize them in order
to make it easier for tooling to discover or update the value cluster
wide.

Refs #22