Leverage buildpack groups
Buildpacks groups are groups of buildpacks that are designed to work
together. Every non-optional buildpack within a group must pass
detection, otherwise, the next group is tested.

This change switches from all function buildpacks being in a single
group with the riff-buildpack acting as a backstop to having a group per
function buildpack.

One minor semantic change is that the first group to detect wins.
Previously, we could enforce that exactly one function buildpack
detected. On the plus side, we no longer need the riff-buildpack.