Make local testing of function buildpacks easier (#50)
The builder.toml is able to accept a local tgz for a buildpack. We can
reduce local development friction to point at local builds for
function buildpacks that are cloned and built on the developers machine.

This is only a convenience for development and should never be used to
ship a builder image.