Pin function buildpack versions
By pinning the version of the function buildpacks we consume, we can
ensure reproducible builds if we need to patch a single buildpack.

The builder.toml file is templated and can be applied by running:

    make templates

The template loads and applies the latest built version. The output of
the template should be committed.

In the future, a CD process could resolve the template and open a PR
when the output changes.