NodeJS Buildpack (#392)
Previously, the builder referred to the NodeJS-related buildpacks
individually.  The NodeJS buildpack team recently made significant changes to
how the buildpacks worked with one another, breaking the builder.  This
changes updates the builder to point at the meta buildpack, which will
insulate us from future changes to the NodeJS buildpack.

Signed-off-by: Ben Hale <bhale@vmware.com>